requirements.txt
fastapi
uvicorn[standard]
jinja2
langchain
langgraph
groq 
python-dotenv
langchain_groq
python-multipart
httpx # For LangChain if making external calls, good to have
